+++ live_sc/test.py
@@ -1,0 +1,130 @@
+from selenium import webdriver
+from selenium.webdriver.common.by import By
+from selenium.webdriver.support.ui import WebDriverWait
+from selenium.webdriver.support import expected_conditions as EC
+from PIL import Image
+from io import BytesIO
+import requests
+import time
+import os
+
+def capture_screenshots(target_string):
+    # Chromeドライバーのパス
+    CHROMEDRIVER_PATH = "./chromedriver.exe"
+
+    # スクリーンショットおよび画像の保存先ディレクトリ
+    SAVE_DIR = "./media/"
+
+    # ライブストリームURLの辞書
+    URLS = {
+        "嵐山": "https://www.youtube.com/watch?v=Op-lf2NRMzs&t=0s",
+        "本願寺": "https://hongwanji-live.securesite.jp/camera/camera01.jpg",
+        "智積院": "https://www.youtube.com/watch?v=n8uPgSSZCEc",
+        "伏見稲荷": "https://www.youtube.com/watch?v=pgqwfZj9HPY",
+        "花見小路": "https://www.youtube.com/watch?v=PXg3ZXgMkGk",
+        "天橋立": "https://www.youtube.com/watch?v=ZCWMo8yzWT0",
+        "ねねの道": "https://www.youtube.com/watch?v=Gxt3YCa2Phc",
+        "北野天満宮": "https://www.youtube.com/watch?v=KHglGodzQ9g",
+        # 他の特定の文字列に対するURLを追加
+    }
+
+    # 対応するURLを取得
+    url = URLS.get(target_string)
+    #print("今は"+url)
+    # もしURLが存在しない場合は終了
+    if url is None:
+        print(f"URL for {target_string} is not found.")
+        return None
+
+    # ブラウザを起動
+    driver = webdriver.Chrome()
+    try:
+        if target_string == "本願寺":
+            # 画像を取得するURL
+            IMAGE_URL = "https://hongwanji-live.securesite.jp/camera/camera01.jpg"
+
+            # 保存先ファイルパスを作成
+            os.makedirs(SAVE_DIR, exist_ok=True)
+            print(IMAGE_URL+"\n")
+            print(url)
+            try:
+                while True:
+                    # 画像を取得
+                    response = requests.get(url)
+                    if response.status_code == 200:
+                        # 画像を保存
+                        timestamp = time.strftime("%Y%m%d%H%M")
+                        filename = f"image_{timestamp}.jpg"
+                        image_path = os.path.join(SAVE_DIR, filename)
+                        with open(image_path, 'wb') as f:
+                            f.write(response.content)
+                        print(f"Image saved: {image_path}")
+
+                    # 10秒待機
+                    time.sleep(10)
+            except KeyboardInterrupt:
+                print("Program terminated by user.")
+        else:
+            screenshots = []  # ここにスクリーンショットを追加していく
+
+            while True:
+                try:
+                    # YouTubeのライブストリームページを開く
+                    driver.get(url)
+
+                    # ウェブページの読み込みが完了するまで待機
+                    wait = WebDriverWait(driver, 20)
+                    wait.until(EC.visibility_of_element_located((By.CLASS_NAME, "ytp-large-play-button")))
+
+                    # 動画再生ボタンがクリック可能になるまで待機
+                    play_button = wait.until(EC.element_to_be_clickable((By.CLASS_NAME, "ytp-large-play-button")))
+                    play_button.click()
+
+                    # 再生部分の要素を特定して位置とサイズを取得
+                    video_element = wait.until(EC.presence_of_element_located((By.CLASS_NAME, "html5-main-video")))
+                    location = video_element.location
+                    size = video_element.size
+
+                    location['y'] += 80
+
+                    count = 1
+                    while True:
+                        try:
+                            # スクリーンショットを取得して保存
+                            screenshot = driver.get_screenshot_as_png()
+                            screenshot = Image.open(BytesIO(screenshot))
+                            screenshot = screenshot.crop((location['x'], location['y'], location['x'] + size['width'], location['y'] + size['height']))
+                            screenshots.append(screenshot)  # リストに追加
+                            print(f"Screenshot {count} taken")
+
+                            # 次のスクリーンショットまで待機
+                            time.sleep(10)  # 10秒ごとにスクリーンショットを取得
+
+                            count += 1
+                        except Exception as e:
+                            print("An error occurred while taking screenshot:", e)
+                            # エラーが発生した場合はページを再読み込みして続行
+                            driver.refresh()
+                except Exception as e:
+                    print("An error occurred:", e)
+                    # エラーが発生した場合はプログラムを継続
+                    continue
+    except KeyboardInterrupt:
+        print("Program terminated by user.")
+    finally:
+        # ブラウザを閉じる
+        driver.quit()
+    
+    return screenshots  # スクリーンショットのリストを返す
+
+# 他のプログラムから呼び出す際に文字列を引数として渡す
+# 嵐山という文字列を引数として渡す
+
+
+# スクリーンショットのリストを読み込む
+screenshots = capture_screenshots("天橋立")
+
+# 各スクリーンショットを表示する
+for idx, screenshot in enumerate(screenshots):
+    screenshot.show()
+
